Refactoring RETS constructor. See extended commit message.
I’ve refactored or added the following to be members instead of variables:

- this.options
- this.config
- this.defaults

I’m doing this so they can be accessed and facilitate debugging but I DO NOT think this pattern should apply to all functions. The idea here is that we want to end up with a config object after applying defaults and options passed from the user. Options still need to be validated though.

I also added some notes, documentation and to-dos.